Renamed Infra/Framework to Infra/Symfony
