+++ apps/api/bin/console
@@ -1,7 +1,7 @@
 #!/usr/bin/env php
 <?php
 
-use Infra\Framework\Kernel;
+use Infra\Symfony\Kernel;
 use Symfony\Bundle\FrameworkBundle\Console\Application;
 
 if (!is_file(dirname(__DIR__).'/vendor/autoload_runtime.php')) {
