+++ apps/api/public/index.php
@@ -1,6 +1,6 @@
 <?php
 
-use Infra\Framework\Kernel;
+use Infra\Symfony\Kernel;
 
 require_once dirname(__DIR__).'/vendor/autoload_runtime.php';
