+++ apps/api/src/Application/HTTP/Controller/Channel.php
@@ -6,7 +6,7 @@ use Application\HTTP\Payload;
 use Domain\Identity\Identifier;
 use Domain\Message\Channel as Message;
 use Domain\Model\User;
-use Infra\Framework\Controller\BaseController;
+use Infra\Symfony\Controller\BaseController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
