+++ apps/api/src/Application/HTTP/Controller/Config.php
@@ -3,7 +3,7 @@
 namespace Application\HTTP\Controller;
 
 use Domain\Message\Config as Message;
-use Infra\Framework\Controller\BaseController;
+use Infra\Symfony\Controller\BaseController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
