+++ apps/api/src/Application/HTTP/Controller/User.php
@@ -5,7 +5,7 @@ namespace Application\HTTP\Controller;
 use Application\HTTP\Payload;
 use Domain\Message\User as Message;
 use Domain\Model\User as UserModel;
-use Infra\Framework\Controller\BaseController;
+use Infra\Symfony\Controller\BaseController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
