+++ apps/api/src/Infra/Symfony/Security/UserProvider/PersistedProvider.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Infra\Framework\Security\UserProvider;
+namespace Infra\Symfony\Security\UserProvider;
 
 use Domain\Model\User;
 use Domain\Repository\Users;
